requirements.txt
pillow==11.2.1
rarfile==4.2
